- renamed pagesDirectory constant to aemPages
